+++ engine/test/mock_app/mock_post_preprocess_no_datatype.py
@@ -15,10 +15,10 @@ async def __init_event__(context: EventContext):
     initialized = True
 
 
-async def __preprocess__(payload: MockData, context: EventContext, request: PreprocessHook,
+async def __preprocess__(payload: None, context: EventContext, request: PreprocessHook,
                          *, query_arg1: str) -> MockData:
-    assert isinstance(payload, MockData)
-    assert request.payload_raw == b'{"value": "ok"}'
+    assert payload is None
+    assert request.payload_raw == b'OK\n'
     return MockData(value=request.headers.get('X-Track-Request-Id'))